Make usage of Global Optimisation effort as initial optional
In very complex mixed fisheries systems, using Global Optimisation effort estimates as the initial values for subsequent optimisation for fully consume 'choke' stock quota can lead to small underutilisation of quota if the optimiser gets stuck in local minima.

The solution is to use uninformative starting values for each stage of optimisation - this is now the default setting. There may be a small increase in computational cost - I have not tested this - but the previous behaviour could be re-enabled via global arguments.